requirements.txt
pandas==2.2.2
rapidfuzz==3.9.6
google-generativeai==0.8.3
python-dotenv==1.0.1
diffusers==0.30.2
transformers==4.44.2
torch==2.3.1
accelerate==0.33.0
pillow==10.4.0
streamlit>=1.39.0
rapidfuzz>=3.9.3
